+++ doc/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix update check.
 - Fix the wrong behavior of the built-in checker Ignore trailing spaces.
 - Remove the misleading "language" in confirmation of submitting to CF.
+- Check whether CF Tool is started when submitting to CF.
 
 ## v6.1